requirements.txt
multidict==6.0.2
yarl==1.8.1
frozenlist==1.3.1
discord.py==2.1.0
ruamel.yaml
schedule
discord-flags
dateparser